requirements.txt
Flask
google-cloud-vision
